requirements.txt
streamlit>=1.32.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyproj>=3.6.0
lxml>=5.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
